PRSD-353: Moves data presentation functions from RegistrationNumberService to RegistrationNumberDataModel
