+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/controllers/ExampleStartPageController.kt
@@ -6,14 +6,11 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import uk.gov.communities.prsdb.webapp.constants.SERVICE_NAME
 import uk.gov.communities.prsdb.webapp.constants.enums.RegistrationNumberType
-import uk.gov.communities.prsdb.webapp.database.entity.RegistrationNumber
-import uk.gov.communities.prsdb.webapp.services.RegistrationNumberService
+import uk.gov.communities.prsdb.webapp.models.dataModels.RegistrationNumberDataModel
 
 @Controller
 @RequestMapping("/")
-class ExampleStartPageController(
-    val regNumService: RegistrationNumberService,
-) {
+class ExampleStartPageController {
     @GetMapping
     fun index(model: Model): String {
         model.addAttribute("contentHeader", "Welcome to the Private Rental Sector Database")
@@ -21,7 +18,7 @@ class ExampleStartPageController(
         model.addAttribute("serviceName", SERVICE_NAME)
         model.addAttribute(
             "landlordRegNum",
-            regNumService.regNumToString(RegistrationNumber(RegistrationNumberType.LANDLORD, 205498766)),
+            RegistrationNumberDataModel(RegistrationNumberType.LANDLORD, 205498766).toString(),
         )
         model.addAttribute("startButtonHref", "/registration")
         model.addAttribute("startButtonText", "Start now")
